% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{Compute the density of a multivariate normal distribution}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{a matrix, with n columns (the observations) and  p rows}

\item{mean}{a vector of means}

\item{varcovM}{a variance-covariance matrix}

\item{Log}{logical, indicating whether the function should return the log of the density? Default is \code{TRUE}}
}
\value{
a list containing the matrix x, and a vector of length n of the multivariate normal distribution density values at those points
}
\description{
mvnpdf() computes the value of the density of any multivariate normal distribution, at n points supplied by the user.
}
\examples{
matrix <- matrix(data = rnorm(100), nrow = 10, ncol = 10)
mean <- apply(matrix, 1, mean)
covmat <- diag(1, 10, 10)
mvnpdf(x = matrix, mean = mean, varcovM = covmat)

}
